Attribute __version__ is added
The attribute __version__ is now accessible in pytim as well
as in setup.py. The file version.py contains the verison information.
When bumping a version, this should be modified only with the git tags.
```python
import pytim
pytim.__version__
```
The __version__ is only included at the top level of the package, the
__init__.py, conforming to the general python style.