Expose `normalize` option to `writecube` (#452)
Currently, pytim normalises the calculated density field by default when writing to a cube. This just exposes the normalize flag to the WillardChandler.writecube() method so you can disable that behaviour -- makes it much easier to find the right values in visualisation software.

This is a redo of #436 with an isolated branch.